+++ core/src/main/java/com/facebook/ktfmt/cli/ParsedArgs.kt
@@ -18,6 +18,7 @@ package com.facebook.ktfmt.cli
 
 import com.facebook.ktfmt.format.Formatter
 import com.facebook.ktfmt.format.FormattingOptions
+import com.facebook.ktfmt.util.Ktfmt
 import java.io.File
 import java.nio.charset.StandardCharsets.UTF_8
 
@@ -70,6 +71,7 @@ data class ParsedArgs(
         |    
         |Commands options:
         |  -h, --help                        Show this help message
+        |  -v, --version                     Show version
         |  -n, --dry-run                     Don't write to files, only report files which 
         |                                        would have changed
         |  --meta-style                      Use 2-space block indenting (default)
@@ -106,6 +108,9 @@ data class ParsedArgs(
       var stdinName: String? = null
 
       if ("--help" in args || "-h" in args) return ParseResult.ShowMessage(HELP_TEXT)
+      if ("--version" in args || "-v" in args) {
+        return ParseResult.ShowMessage("ktfmt version ${Ktfmt.version}")
+      }
 
       for (arg in args) {
         when {
@@ -119,7 +124,8 @@ data class ParsedArgs(
               stdinName =
                   parseKeyValueArg("--stdin-name", arg)
                       ?: return ParseResult.Error(
-                          "Found option '${arg}', expected '${"--stdin-name"}=<value>'")
+                          "Found option '${arg}', expected '${"--stdin-name"}=<value>'"
+                      )
           arg.startsWith("--") -> return ParseResult.Error("Unexpected option: $arg")
           arg.startsWith("@") -> return ParseResult.Error("Unexpected option: $arg")
           else -> fileNames.add(arg)
@@ -132,7 +138,8 @@ data class ParsedArgs(
           val filesExceptStdin = fileNames - "-"
           return ParseResult.Error(
               "Cannot read from stdin and files in same run. Found stdin specifier '-'" +
-                  " and files ${filesExceptStdin.joinToString(", ")} ")
+                  " and files ${filesExceptStdin.joinToString(", ")} "
+          )
         }
       } else if (stdinName != null) {
         return ParseResult.Error("--stdin-name can only be specified when reading from stdin")
@@ -145,7 +152,8 @@ data class ParsedArgs(
               dryRun,
               setExitIfChanged,
               stdinName,
-          ))
+          )
+      )
     }
 
     private fun parseKeyValueArg(key: String, arg: String): String? {